refactor: move classes into session packge
